Just move some stuff to obj
