+++ wasm/src/dbg/mod.rs
@@ -1,4 +1,5 @@
 pub mod compopts;
 pub mod handler;
+pub mod obj;
 pub mod server;
 pub mod types;
